+++ src/onegov.form/onegov/form/models/__init__.py
@@ -3,9 +3,11 @@ from onegov.form.models.submission import FormSubmission
 from onegov.form.models.submission import PendingFormSubmission
 from onegov.form.models.submission import CompleteFormSubmission
 from onegov.form.models.submission import FormFile
+from onegov.form.models.registration_window import FormRegistrationWindow
 
 __all__ = (
     'FormDefinition',
+    'FormRegistrationWindow',
     'FormSubmission',
     'PendingFormSubmission',
     'CompleteFormSubmission',