+++ src/onegov.form/onegov/form/upgrade.py
@@ -5,14 +5,14 @@ upgraded on the server. See :class:`onegov.core.upgrade.upgrade_task`.
 from depot.io.utils import FileIntent
 from io import BytesIO
 from onegov.core.crypto import random_token
-from onegov.core.orm.types import JSON
+from onegov.core.orm.types import JSON, UUID
 from onegov.core.upgrade import upgrade_task
 from onegov.core.utils import dictionary_to_binary
 from onegov.core.utils import normalize_for_url
 from onegov.form import FormDefinitionCollection
 from onegov.form import FormFile
 from onegov.form import FormSubmission
-from sqlalchemy import Column, Text
+from sqlalchemy import Column, Integer, Text
 
 
 @upgrade_task('Enable external form submissions')
@@ -122,4 +122,23 @@ def add_group_order_to_form_definitions(context):
         table='forms',
         column=Column('order', Text, nullable=False, index=True),
         default=lambda form: normalize_for_url(form.title)
+    )
+
+
+@upgrade_task('Add registration window columns')
+def add_registration_window_columns(context):
+    context.operations.add_column(
+        'submissions',
+        Column('claimed', Integer, nullable=True)
+    )
+
+    context.operations.add_column(
+        'submissions',
+        Column('registration_window_id', UUID, nullable=True)
+    )
+
+    context.add_column_with_defaults(
+        table='submissions',
+        column=Column('spots', Integer, nullable=False),
+        default=0
     )